requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
Django==4.0.3
djangorestframework==3.13.1
google-auth==2.6.2
idna==3.3
jsonfield==3.1.0
kubernetes==23.3.0
loguru==0.6.0
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
websocket-client==1.3.1
win32-setctime==1.1.0
